Modularise macos wrappers files
