+++ src/Makefile
@@ -1017,8 +1017,11 @@ wrappers_libc_o := $(wrappers_libc_c:.c=.o)
 wrappers_linux_c := $(wildcard unit_tests/wrappers/linux/*.c)
 wrappers_linux_o := $(wrappers_linux_c:.c=.o)
 
-wrappers_macos_c := $(wildcard unit_tests/wrappers/macos/libc/*.c)
+wrappers_macos_c := $(wildcard unit_tests/wrappers/macos/*.c)
 wrappers_macos_o := $(wrappers_macos_c:.c=.o)
+
+wrappers_macos_libc_c := $(wildcard unit_tests/wrappers/macos/libc/*.c)
+wrappers_macos_libc_o := $(wrappers_macos_libc_c:.c=.o)
 
 wrappers_posix_c := $(wildcard unit_tests/wrappers/posix/*.c)
 wrappers_posix_o := $(wrappers_posix_c:.c=.o)
@@ -1099,6 +1102,7 @@ ifneq (,$(filter ${TEST},YES yes y Y 1))
 		UNIT_TEST_WRAPPERS+=${wrappers_windows_posix_o}
 	else ifeq (${uname_S},Darwin)
 	    UNIT_TEST_WRAPPERS+=${wrappers_macos_o}
+		UNIT_TEST_WRAPPERS+=${wrappers_macos_libc_o}
 	else
 		UNIT_TEST_WRAPPERS+=${wrappers_externals_audit_o}
 		UNIT_TEST_WRAPPERS+=${wrappers_externals_procpc_o}
@@ -2005,6 +2009,7 @@ clean-unit-tests:
 	rm -f ${wrappers_libc_o}
 	rm -f ${wrappers_linux_o}
 	rm -f ${wrappers_macos_o}
+	rm -f ${wrappers_macos_libc_o}
 	rm -f ${wrappers_posix_o}
 	rm -f ${wrappers_wazuh_o}
 	rm -f ${wrappers_wazuh_os_crypto_o}
