+++ src/unit_tests/wrappers/macos/libplist_wrappers.h
@@ -1,0 +1,17 @@
+#ifndef STDIO_WRAPPERS_LIBPLIST_H
+#define STDIO_WRAPPERS_LIBPLIST_H
+
+#include <plist/plist.h>
+
+#undef plist_from_bin
+#define plist_from_bin wrap_plist_from_bin
+#undef plist_to_xml
+#define plist_to_xml wrap_plist_to_xml
+#undef plist_free
+#define plist_free wrap_plist_free
+
+void wrap_plist_from_bin (char * bin, size_t size, plist_t *node);
+void wrap_plist_to_xml (plist_t *node, char ** xml, uint32_t *size);
+void wrap_plist_free(plist_t node);
+
+#endif
